// src/app/core/models/facturation/expedition-item.model.ts
export class ExpeditionItem {
    created_at: string;
    declared_value: string;
    height: string;
    id: number;
    id_expedition: number;
    item_comment: string | null;
    length: string;
    nature: string;
    number: number;
    product_category_id: number | null;
    ticket_grouped: boolean;
    type: string;
    updated_at: string;
    uuid: string;
    volume: string;
    volumetric_weight: number | null;
    weight: string;
    width: string;
}